Score - Singleton
Implementation of an account using the Singleton pattern

+++ ArkanoidGame/LevelLoader.cpp
@@ -58,6 +58,6 @@ namespace ArkanoidGame {
 
 	int LevelLoader::GetLevelCount()
 	{
-		return levels.size();
+		return (int)levels.size();
 	}
 }
